+++ NEWS.md
@@ -1,5 +1,6 @@
 # pastapi 0.1.0.9000
 
+* Simpler testing
 
 # pastapi 0.1.0
 
@@ -7,5 +8,5 @@
 
 # pastapi 0.0.0.9000
 
-* Initial commit to github.
-* Added a `NEWS.md` file to track changes to the package.
+* Initial commit to github
+* Added a `NEWS.md` file to track changes to the package